# apps/cockpit/openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityIn": {
        "properties": {
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "key",
          "status"
        ],
        "title": "CapabilityIn",
        "type": "object"
      },
...
      "EnforcementApplyIn": {
        "properties": {
          "environment": {
            "title": "Environment",
            "type": "string"
          }
        },
        "required": [
          "environment"
        ],
        "title": "EnforcementApplyIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DQ & Observability Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/enforcement/apply": {
      "post": {
        "description": "Bootstrap der Gate-Infrastruktur anwenden. DDL im Tenant-Schema ist\neine Betriebsentscheidung \u2014 owner/admin; jede Anwendung ist eine\nauditierte Operation.",
        "operationId": "apply_plan_api_enforcement_apply_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnforcementApplyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Plan",
        "tags": [
          "enforcement"
        ]
      }
    },
    "/api/enforcement/capabilities": {
      "get": {
        "description": "Verifizierte Tenant-F\u00e4higkeiten (Rest-O5/O6) + offene manuelle Checks.\nGrundlage f\u00fcr Aktivierungs-Entscheidungen der Slices \u2463/\u2465.",
        "operationId": "list_capabilities_api_enforcement_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "enforcement"
        ]
      },
      "post": {
        "description": "Ergebnis eines manuellen Spike-Checks eintragen (Spike-Kit \u00a7Ablauf).",
        "operationId": "record_capability_api_enforcement_capabilities_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Capability",
        "tags": [
          "enforcement"
        ]
      }
    },
    "/api/enforcement/plan": {
      "get": {
        "description": "Soll-Zustand des Signal-Schemas (Dry-Run): globale Infrastruktur\n(inkl. Bridge-Prozeduren bei Opt-in) und Split-Artefakte je Objekt\n(Slice \u2463/\u2464). Zeigt DDL \u2014 steward+.",
        "operationId": "get_plan_api_enforcement_plan_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "enforcement"
        ]
      }
    },
    "/api/enforcement/probe": {
      "post": {
        "description": "Automatisierbare Pre-Flight-Probes am Tenant ausf\u00fchren (harmlose\nProbe-Objekte im eigenen Schema, sofort gedroppt). Bewusst OHNE den\nMaterialisierungs-Kill-Switch nutzbar \u2014 die Probe ist der Schritt DAVOR;\nnur das Ziel-Schema muss konfiguriert sein.",
        "operationId": "run_probe_api_enforcement_probe_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnforcementApplyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Probe",
        "tags": [
          "enforcement"
        ]
      }
    },
...
    "/api/quarantine": {
      "get": {
        "operationId": "list_quarantine_api_quarantine_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Quarantine",
        "tags": [
          "quarantine"
        ]
      }
    },
    "/api/quarantine/{quarantine_id}": {
      "get": {
        "operationId": "get_quarantine_api_quarantine__quarantine_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "quarantine_id",
            "required": true,
            "schema": {
              "title": "Quarantine Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quarantine",
        "tags": [
          "quarantine"
        ]
      }
    },
...
  }
}